pipeline/requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
python-slugify>=8.0.0
google-genai>=1.0.0
jsonschema>=4.21.0
SPARQLWrapper>=2.0.0
